+++ dsm.py
@@ -4,17 +4,6 @@ Direct Stifness Method
 Mauro S. Maza - FCEFyN (UNC) - 2024-06-23
 """
 
-"""
-COSAS POR IMPLEMENTAR
-- ElmFrame2D: incorporar la opción para articular extremos
-- Node.add_dofs: asegurarse que el arreglo quede como vector columna    
-        # IMPLEMENTAR MÉTODO PARA AGREGAR MUCHOS NODOS Y MUCHOS ELEMENTOS A LA VEZ
-        # IMPLEMENTAR MÉTODO PARA ACTUALIZAR COORDENADAS DE UN PUNTO
-        # PROBAR SI AL CAMBIAR LAS COORDENADA DE UN NODO, SE ACTUALIZA LA REFERENCIA EN EL ELEMENTO
-        
-        # se puede mejorar la programación creando una función que ensamble cualquier cosa y usando eso múltiples veces
-"""
-
 # IMPORTING ZONE
 import numpy as np
 from typing import List
@@ -591,40 +580,87 @@ class Struc2D3Dof:
 # RUN
 if __name__ == '__main__':
     
-    # simple structure
-    if False:
-        # create structure object
+    """
+    Example #1 for dsm.py
+    RC Hibbeler - Análisis Estructural - 8ed - Ejemplo 16.1
+    In SI units
+    Origin of coord. sys. on the bottom-left corner
+    """
+    if True:
+        # import numpy as np
+        # from dsm import Struc2D3Dof
+
+        # conversion factors
+        # input
+        feet2mm = 304.8
+        ksi2MPa = 6.89476
+        in22mm2 = 645.16
+        in42mm4 = 416231.426
+        kip2N   = 4448.22
+        # output
+        mm2in   = 0.0393701
+
+        # creat problem
         str=Struc2D3Dof('simple frame')
-        
-        # create node and elem objects and add them to the structure
-        if True: # in one step - recomended
-            str.create_node(10, np.array([0,0  ], dtype=float))
-            str.create_node(20, np.array([0,5.5], dtype=float))
-            str.create_node(30, np.array([3,0  ], dtype=float))
-            str.create_node(40, np.array([3,5.5], dtype=float))
-            
-            str.create_elm(1, [10,20], 210e3, 100, 10**4/12)
-            str.create_elm(2, [10,40], 210e3, 100, 10**4/12)
-            str.create_elm(3, [20,40], 210e3, 100, 10**4/12)
-            
-        else: # in two steps - not recomended
-            n1=Node(10, np.array([0,0  ], dtype=float))
-            n2=Node(20, np.array([0,5.5], dtype=float))
-            n3=Node(30, np.array([3,0  ], dtype=float))
-            n4=Node(40, np.array([3,5.5], dtype=float))
-            str.add_node(n1)
-            str.add_node(n2)
-            str.add_node(n3)
-            str.add_node(n4)
-        
-            elm1=ElmFrame2D(1, [n1,n2], 210e3, 100, 10**4/12)
-            elm2=ElmFrame2D(2, [n1,n4], 210e3, 100, 10**4/12)
-            elm3=ElmFrame2D(3, [n2,n4], 210e3, 100, 10**4/12, joints='both')
-            str.add_elm(elm1)
-            str.add_elm(elm2)
-            str.add_elm(elm3)
-    
-    
-    
+
+        # nodes
+        str.create_node(1, np.array([ 0*feet2mm, 20*feet2mm], dtype=float))
+        str.create_node(2, np.array([20*feet2mm, 20*feet2mm], dtype=float))
+        str.create_node(3, np.array([20*feet2mm,  0*feet2mm], dtype=float))
+
+        # elms
+        E=29e3*ksi2MPa
+        A=10*in22mm2
+        I=500*in42mm4
+        str.create_elm(1, [1,2],  E, A, I)
+        str.create_elm(2, [2,3],  E, A, I)
+
+        # loads
+        str.create_NL(2, 0, 5*kip2N)
+
+        # BCs
+        str.create_BC(1, 1, 0)
+        str.create_BC(3, 0, 0)
+        str.create_BC(3, 1, 0)
+        str.create_BC(3, 2, 0)
+
+        # solve
+        str.assemble()
+        str.impose_BCs()
+        str.solve()
+
+        # print
+        print( 'Unknwon displacements:')
+        print(f'> D1: node 2, u_x = {str.nodes[1].dofVal[0]*mm2in} in')
+        print(f'> D2: node 2, u_y = {str.nodes[1].dofVal[1]*mm2in} in')
+        print(f'> D3: node 2, phi = {str.nodes[1].dofVal[2]} rad')
+        print(f'> D4: node 1, u_x = {str.nodes[0].dofVal[0]*mm2in} in')
+        print(f'> D5: node 1, phi = {str.nodes[0].dofVal[2]} rad')
+        print( '')
+        print( 'Reactions:')
+        idx=np.where( str.reacts.row == str.nodes[0].dofLab[1] )[0][0]
+        print(f'> Q6: node 1, R_y = { str.reacts.vtr[idx]/kip2N } kip')
+        idx=np.where( str.reacts.row == str.nodes[2].dofLab[0] )[0][0]
+        print(f'> Q7: node 3, R_x = { str.reacts.vtr[idx]/kip2N } kip')
+        idx=np.where( str.reacts.row == str.nodes[2].dofLab[1] )[0][0]
+        print(f'> Q8: node 3, R_y = { str.reacts.vtr[idx]/kip2N } kip')
+        idx=np.where( str.reacts.row == str.nodes[2].dofLab[2] )[0][0]
+        print(f'> Q9: node 3, Mom = { str.reacts.vtr[idx]/kip2N*mm2in } kip*in')
+        print( '')
+        print( 'Elemental nodal loads in global coordinates:')
+        print( 'element #1')
+        print(f'> elem 1, node i, L_xg = {str.elmts[0].L_glo[0]/kip2N} kip')
+        print(f'> elem 1, node i, L_yg = {str.elmts[0].L_glo[1]/kip2N} kip')
+        print(f'> elem 1, node i, mome = {str.elmts[0].L_glo[2]/kip2N*mm2in} kip*in')
+        print(f'> elem 1, node j, L_xg = {str.elmts[0].L_glo[3]/kip2N} kip')
+        print(f'> elem 1, node j, L_yg = {str.elmts[0].L_glo[4]/kip2N} kip')
+        print(f'> elem 1, node j, mome = {str.elmts[0].L_glo[5]/kip2N*mm2in} kip*in')
+        print( 'element #2')
+        print(f'> elem 2, node i, L_xg = {str.elmts[1].L_glo[0]/kip2N} kip')
+        print(f'> elem 2, node i, L_yg = {str.elmts[1].L_glo[1]/kip2N} kip')
+        print(f'> elem 2, node i, mome = {str.elmts[1].L_glo[2]/kip2N*mm2in} kip*in')
+        print(f'> elem 2, node j, L_xg = {str.elmts[1].L_glo[3]/kip2N} kip')
+        print(f'> elem 2, node j, L_yg = {str.elmts[1].L_glo[4]/kip2N} kip')
+        print(f'> elem 2, node j, mome = {str.elmts[1].L_glo[5]/kip2N*mm2in} kip*in')